Sort threads by last message date in document ThreadManager queries
Due to MongoDB limitations and the current schema, we cannot sort by the timestamps present in the metadata fields and match behavior of the ORM's ThreadManager. This would require use of Mongo's positional operator ($) in sorts, but it may be possible with the upcoming aggregation framework in MongoDB 2.2. TODO's have been left to revisit this when possible.

To compromise, we can sort by the creation timestamp of the last message in the thread, which is easy to denormalize and index.