Replace metadata document queries with indexable array fields
MongoDB does not currently take advantage of multi-key indexes for the $elemMatch operator. This is planned as a future improvement. See: https://jira.mongodb.org/browse/SERVER-3104

As a result, the previous indexes and $elemMatch queries were not efficient for large message collections. This refactoring maintains the metadata arrays, but introduces array fields specifically tailored to the following queries:

 * thread inbox
 * thread sentbox
 * thread searching (note: the keyword regex is still inefficient)
 * counting unread messages

These new fields are maintained via denormalization routines, and they are now created as part of the database migration script.