Update Resources/config/routing.xml
I did it because Symfony2.1 load fos_message_thread_view in priority of the other path when we try to delete a thread. With this modification, the path fos_message_thread_delete works because the routing service load it before. It was conflictual because this two paths begins with same text.