Fixes interface breaking method signature in MessageManager.
This will fix the errors introduced in pull request #89. Furthermore it fixes issue #91 opened because of #89.
Originally there was the issue #87 which brought not only an inconsistency but an error to my attention. Now the methods markAsReadByParticipant and markAsUnreadByParticipant forward the calls correctly to the methods in the given readable (like the interface ReadableInterface indicates).